http://bugs.gentoo.org/show_bug.cgi?id=23574 http://bugs.gentoo.org/show_bug.cgi?id=23518
Add warning on top of "Booting" to warn that ignoring the bootoptions
could potentially result in wrong keyboard settings, unstarted pcmcia
services and more.

Inform the users about the dangers of root, add section to inform the user
more precisely that he should be adding a user for day-to-day tasks.