A few updates to the policy, sufficient enough to do a major version update.
First of all, the need for translators to start a bugreport on bugs.gentoo.org
when they start a new translation has been lifted. As the lead translator is in
charge, (s)he should be given the opportunity to manage and develop the
translations as (s)he sees fit. This is already done in practice, so it wasn't a
real policy.

Second of all, reviewing documents has been worn off. Why? Because we have too
little resources to keep asking for reviews of updates that are trivial or of
which the editor is sure the changes are correct. The new policy states that
updates can be propagated to CVS immediately _unless_ the editor is not 100%
confident of the updates, in which case a review of a Gentoo developer is
requested.

Third, high-volume updates, technical updates or intrusive updates must be
accompanied by a bugreport on bugs.gentoo.org of which the bugnumber _must_ be
mentioned in the CVS log. This allows for backtracking changes to the
documentation (which has been proven to be *very* useful :)

For those of you reading this from the cvs-doc mailinglist, I'll post the exact
message on gentoo-doc in a minute, just be patient :p