#87117 - Add "udev" as boot parameter
Apparently, the issues aren't always fixed. I have both reports of success with
the current genkernel (3.1.6) and malfunctions. Again, the issue seems to be
with the initrd created by genkernel, and not with the genkernel that's on the
Universal InstallCD (networkless installations aren't affected).

We'll probably want to remove this update from the guide later on when we're
certain that it's fixed.