Deprecated project page. The Gentoo tools repository can be found here:
http://gentooexperimental.org/script/repo/list

None of these tools tend to have their own webpage (actually only mailmin has a
project page), so it is a bit overkill to keep the tools page. Cfr. the Fosdem
metastructure reform proposal for a better way to organize this on gentoo
infrastructure. It is probably a good idea to use a wiki for the pages of these
tools since they change quite frequently.