m ravilj@gmail.com  Mon Aug  1 11:06:01 2005 Date: Sun, 31 Jul 2005 23:06:22 +0200 From: Ryan Viljoen <ravilj@gmail.com> To: www@gentoo.org Subject: Gentoo Badge for Forum Signitures
[-- Attachment #1 --]
[-- Type: text/plain, Encoding: quoted-printable, Size: 0.4K --]

Hi

I noticed that a lot of people are using small badges in their
signitures on forums like in the attached picture. The firefox spread
firefox badge is also widely used. So figured why not a Gentoo one so
I created a Gentoo one for my signiture:
http://babbage.ee.wits.ac.za/~viljoenr/gentoo.png

Anyway feel free to make use of it or add it to the Graphics and logo
page or whatever.

Cheers