Introduce version attribute for the license tag in GuideXML
The license tag is used to have a document marked as being under the CC-BY-SA license. Currently, it holds the function of
refering to the CC-BY-SA 2.5 version. However, the license itself has evolved and is currently on the 3.0 version.
Documents created in the Gentoo wiki (wiki.gentoo.org) automatically get the 3.0 license.

With the version attribute, we can now enter a specific version for the license while keeping the old behavior at hand (as
we are not allowed to switch licenses on existing material without prior agreement of the copyright holders and authors.

This commit is the first of many. The DTD change should happen before any other as the DTD files are also distributed
elsewhere.