get_glsas.py: Rewrite get_glsas.py from scratch
The old one didn't work and I had not idea how to make it work.
Rewrite it to use the glsa table from
http://www.gentoo.org/security/en/glsa/index.xml?passthrough=1
and BeautifulSoup (http://www.crummy.com/software/BeautifulSoup/)
to make it work again. Print text compatible for direct inclusion in the
WP GMN. The code is not ideal but it works ;)